Rename variables n, n2, ..., and max_len (see full message)
Rename:
- n to NN
- n2 to N2, etc
- max_len to n

Having all-cap non constant variable names is not ideal. However, it was
worse previously, imo. If we want to change later, it will however be
much easier, as there are a lot of words containing n, but NN is
uniquely our variable.

Note that NN should actually be named N', but that's not really
possible. N2 could be used both in math and code, but that would
contradict the definitions of N3, etc (which could be renamed to N_3
etc, tho).